Multiple fit parameters functionality
It is now possible to specify lists of t_start_fit and t_end_fit in
input, instead of simple integers. The code will compute the MSD once, and
a linear regression for each (t_start_fit, t_end_fit) pair.
Lists of slopes/diffusions will be returned in the results dictionary.
This functionality allows one to study the convergence of the diffusion as
a function of the time-window used for fitting, at a reduced cost.